Refactor out RenderingUtils
