+++ src/main/java/com/glencoesoftware/omero/ms/image/region/ShapeMaskCtx.java
@@ -18,6 +18,8 @@
 
 package com.glencoesoftware.omero.ms.image.region;
 
+import java.util.Arrays;
+import java.util.List;
 import java.util.Optional;
 
 import org.slf4j.LoggerFactory;
@@ -26,6 +28,7 @@ import com.glencoesoftware.omero.ms.core.OmeroRequestCtx;
 
 import io.vertx.core.MultiMap;
 import ome.model.roi.Mask;
+import ome.xml.model.primitives.Color;
 import omeis.providers.re.data.RegionDef;
 
 public class ShapeMaskCtx extends OmeroRequestCtx {
@@ -40,7 +43,7 @@ public class ShapeMaskCtx extends OmeroRequestCtx {
     public Long shapeId;
 
     /** Display color */
-    public String color;
+    public Color color;
 
     /** Whether or not to flip horizontally */
     public boolean flipHorizontal;
@@ -73,7 +76,7 @@ public class ShapeMaskCtx extends OmeroRequestCtx {
         try {
             this.omeroSessionKey = omeroSessionKey;
             shapeId = Long.parseLong(params.get("shapeId"));
-            color = params.get("color");
+            color = splitHTMLColor(params.get("color"));
             String flip = Optional.ofNullable(params.get("flip"))
                     .orElse("").toLowerCase();
             flipHorizontal = flip.contains("h");
@@ -116,5 +119,42 @@ public class ShapeMaskCtx extends OmeroRequestCtx {
     public String cacheKey() {
         return String.format(
                 CACHE_KEY_FORMAT, Mask.class.getName(), shapeId, color);
+    }
+
+    /**
+     *  Splits an hex stream of characters into an array of bytes
+     *  in format (R,G,B,A) and converts to a
+     *  {@link ome.xml.model.primitives.Color}.
+     *  - abc      -> (0xAA, 0xBB, 0xCC, 0xFF)
+     *  - abcd     -> (0xAA, 0xBB, 0xCC, 0xDD)
+     *  - abbccd   -> (0xAB, 0xBC, 0xCD, 0xFF)
+     *  - abbccdde -> (0xAB, 0xBC, 0xCD, 0xDE)
+     *  @param color Characters to split.
+     *  @return corresponding {@link ome.xml.model.primitives.Color}
+     */
+    public static Color splitHTMLColor(String color) {
+        List<Integer> level1 = Arrays.asList(3, 4);
+        try {
+            if (level1.contains(color.length())) {
+                String c = color;
+                color = "";
+                for (char ch : c.toCharArray()) {
+                    color += ch + ch;
+                }
+            }
+            if (color.length() == 6) {
+                color += "FF";
+            }
+            if (color.length() == 8) {
+                int r = Integer.parseInt(color.substring(0, 2), 16);
+                int g = Integer.parseInt(color.substring(2, 4), 16);
+                int b = Integer.parseInt(color.substring(4, 6), 16);
+                int a = Integer.parseInt(color.substring(6, 8), 16);
+                return new Color(r, g, b, a);
+            }
+        } catch (Exception e) {
+            log.error("Error while parsing color: {}", color, e);
+        }
+        return null;
     }
 }
