+++ src/main/java/com/glencoesoftware/omero/ms/image/region/ShapeMaskRequestHandler.java
@@ -109,9 +109,7 @@ public class ShapeMaskRequestHandler {
             if (shapeMaskCtx.color != null) {
                 // Color came from the request so we override the default
                 // color the mask was assigned.
-                int[] rgba = RenderingUtils
-                        .splitHTMLColor(shapeMaskCtx.color);
-                fillColor = new Color(rgba[0], rgba[1], rgba[2], rgba[3]);
+                fillColor = shapeMaskCtx.color;
             }
             log.debug(
                 "Fill color Red:{} Green:{} Blue:{} Alpha:{}",
